Data.Bag

Documentation

data Bag a

Datatype for Bag

emptyBag :: Bag a

Empty Bag of type a. O(1)

singleton :: a -> Bag a

Bag containing one element of type a. O(1)

fromList :: Ord a => [a] -> Bag a

Function to create a Bag from a List. O(m lg m + n), where m = # of distinct elements in the List, and n = # of elements in the List

toList :: Bag a -> [a]

Function to convert a Bag into a List. O(n)

add :: Ord a => a -> Bag a -> Bag a

Function to add one element to a Bag of the same type. O(lg n)

remove :: Ord a => a -> Bag a -> Bag a

Function to remove one occurence of an element from the Bag. O(lg n)

member :: Ord a => a -> Bag a -> Bool

Function to check if an element is present in the Bag or not. Use member in place of elem. O(lg n)

count :: (Eq a, Ord a) => a -> Bag a -> Int

Function to count the occurence of an element in the Bag. O(lg n)

unique :: Ord a => Bag a -> Set a

Function to return unique elements from a Bag. O(# of distinct elements)